Creates initial configurations for bar & workspace
This commit creats the initial setup for bar and workspaces.
It is split into three sections, left middle and right.
Currently workspaces is setup, with a date module on the right side.
Workspace hooks into EWMH to know when to update, then calls
bspc queries to actually understand the update.